+++ CHANGES.rst
@@ -4,7 +4,7 @@
 
 New Features
 ^^^^^^^^^^^^
-
+- added feature to specify extension when calling ImageFileCollection [#463]
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^